Move implementations from headers into source files
